server/mcp_server_deploy/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.3
    # via fastmcp
cachetools==6.2.0
    # via mcp-server-deploy (pyproject.toml)
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   volcengine-python-sdk
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
cryptography==45.0.7
    # via authlib
cyclopts==3.24.0
    # via fastmcp
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22
    # via rich-rst
email-validator==2.3.0
    # via pydantic
exceptiongroup==1.3.0
    # via fastmcp
fastmcp==2.12.2
    # via mcp-server-deploy (pyproject.toml)
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
isodate==0.7.2
    # via openapi-core
jsonschema==4.25.0
    # via
    #   mcp
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.12.0
    # via openapi-spec-validator
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via werkzeug
mcp==1.14.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.8.0
    # via openapi-core
mysql-connector-python==9.4.0
    # via mcp-server-deploy (pyproject.toml)
openapi-core==0.19.5
    # via fastmcp
openapi-pydantic==0.5.1
    # via fastmcp
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pycparser==2.23
    # via cffi
pydantic==2.11.7
    # via
    #   fastmcp
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
pyperclip==1.9.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via volcengine-python-sdk
python-dotenv==1.1.1
    # via
    #   mcp-server-deploy (pyproject.toml)
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pyyaml==6.0.2
    # via jsonschema-path
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.5
    # via jsonschema-path
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.1
    # via cyclopts
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
    #   volcengine-python-sdk
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.47.2
    # via mcp
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   openapi-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   volcengine-python-sdk
uvicorn==0.35.0
    # via mcp
volcengine-python-sdk==4.0.20
    # via mcp-server-deploy (pyproject.toml)
werkzeug==3.1.1
    # via openapi-core
